Reduce complexity of feed preference code
