+++ ui/preferences/src/main/res/values/arrays.xml
@@ -7,10 +7,11 @@
         <item>@string/feed_auto_download_never</item>
     </string-array>
 
+    <!-- Keep consistent with AutoDeleteAction -->
     <string-array name="spnAutoDeleteValues">
-        <item>global</item>
-        <item>always</item>
-        <item>never</item>
+        <item>0</item>
+        <item>1</item>
+        <item>2</item>
     </string-array>
 
     <string-array name="spnEnableAutoDownloadItems">
@@ -19,10 +20,11 @@
         <item>@string/disabled</item>
     </string-array>
 
+    <!-- Keep consistent with AutoDownloadSetting -->
     <string-array name="spnEnableAutoDownloadValues">
-        <item>global</item>
-        <item>enabled</item>
-        <item>disabled</item>
+        <item>1</item>
+        <item>2</item>
+        <item>0</item>
     </string-array>
 
     <string-array name="spnVolumeAdaptationItems">
@@ -34,13 +36,14 @@
         <item>@string/feed_volume_boost_heavy</item>
     </string-array>
 
+    <!-- Keep consistent with VolumeAdaptionSetting -->
     <string-array name="spnVolumeAdaptationValues">
-        <item>heavy</item>
-        <item>light</item>
-        <item>off</item>
-        <item>light_boost</item>
-        <item>medium_boost</item>
-        <item>heavy_boost</item>
+        <item>2</item>
+        <item>1</item>
+        <item>0</item>
+        <item>3</item>
+        <item>4</item>
+        <item>5</item>
     </string-array>
 
     <string-array name="feed_refresh_interval_entries">
